requirements.txt
streamlit>=1.32.0

# LLM & orchestration
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.7
langsmith>=0.1.30
openai>=1.30.0

# File parsing
pypdf>=4.2.0
python-docx>=1.1.0

# URL parsing
requests>=2.31.0
beautifulsoup4>=4.12.3

# Utilities
python-dotenv>=1.0.1
